+++ Client/src/Components/Users.jsx
@@ -1,52 +1,23 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
-import { DataGrid } from '@mui/x-data-grid';
-import { Button } from '@mui/material';
-import Header from './Header'; // Make sure you import the header
-import SideBar from './SideBar'; 
+import React from 'react';
+import UsersTable from './UsersTable';
 import './css/users.css';
+import Header from './Header';
+import SideBar from './SideBar';
+
+
 
 const Users = () => {
-  const [users, setUsers] = useState([]);
-
-  useEffect(() => {
-    const fetchUsers = async () => {
-      try {
-        const response = await axios.get('/api/users');
-        setUsers(response.data);
-      } catch (error) {
-        console.error('Failed to fetch users:', error);
-      }
-    };
-
-    fetchUsers();
-  }, []);
-
-  const columns = [
-    { field: 'id', headerName: 'ID', flex: 0.5 },
-    { field: 'name', headerName: 'Name', flex: 1 },
-    { field: 'cardID', headerName: 'Card ID', flex: 1 },
-    { field: 'balance', headerName: 'Balance', flex: 1 },
-    { field: 'actions', headerName: 'Actions', flex: 1, renderCell: (params) => (
-        <div className="actions-cell">
-          <Button variant="contained" color="success" size="small" style={{ marginRight: 8 }}>Edit</Button>
-          <Button variant="contained" color="error" size="small">Delete</Button>
-        </div>
-      )
-    },
-  ];
-
   return (
-    <div style={{ height: 400, width: '100%' }}>
-      <DataGrid
-        rows={users.map((user, index) => ({ ...user, id: index + 1 }))}
-        columns={columns}
-        pageSize={5}
-        rowsPerPageOptions={[5]}
-        disableSelectionOnClick
-        autoHeight
-      />
-    </div>
+    <>
+      <Header />
+      <div className="app-layout">
+        <SideBar />
+        <div className="main-content">
+          <h1>Users</h1>
+          <UsersTable />
+        </div>
+      </div>
+    </>
   );
 };
